refactor: move environment configuration to src/config and update imports
