+++ src/email/mailer/mailer.provider.ts
@@ -1,6 +1,6 @@
 import * as Mailchimp from '@mailchimp/mailchimp_transactional';
 import { Logger } from '@nestjs/common';
-import { envs } from 'config';
+import { envs } from 'src/config';
 
 export const MailerProvider = [
   {
