+++ src/sms/masiva/masiva.service.ts
@@ -1,9 +1,9 @@
 import { HttpService } from '@nestjs/axios';
 import { HttpStatus, Injectable } from '@nestjs/common';
 import { RpcException } from '@nestjs/microservices';
-import { envs } from 'config';
 import * as moment from 'moment';
 import { lastValueFrom } from 'rxjs';
+import { envs } from 'src/config';
 import { DatabaseService } from 'src/database/database.service';
 @Injectable()
 export class MasivaService {
